requirements.txt
Flask
numpy
scipy
pydub
pycryptodome
